Revert "Collect input_SET children in instance."
This reverts commit 052e65ca1befb19049ee9f02f472d20cf78d8dc1.
