+++ openpype/hosts/maya/api/lib.py
@@ -3562,43 +3562,3 @@ def get_color_management_output_transform():
     if preferences["output_transform_enabled"]:
         colorspace = preferences["output_transform"]
     return colorspace
-
-
-def get_all_children(nodes):
-    """Return all children of `nodes` including each instanced child.
-    Using maya.cmds.listRelatives(allDescendents=True) includes only the first
-    instance. As such, this function acts as an optimal replacement with a
-    focus on a fast query.
-
-    """
-
-    sel = OpenMaya.MSelectionList()
-    traversed = set()
-    iterator = OpenMaya.MItDag(OpenMaya.MItDag.kDepthFirst)
-    for node in nodes:
-
-        if node in traversed:
-            # Ignore if already processed as a child
-            # before
-            continue
-
-        sel.clear()
-        sel.add(node)
-        dag = sel.getDagPath(0)
-
-        iterator.reset(dag)
-        # ignore self
-        iterator.next()  # noqa: B305
-        while not iterator.isDone():
-
-            path = iterator.fullPathName()
-
-            if path in traversed:
-                iterator.prune()
-                iterator.next()  # noqa: B305
-                continue
-
-            traversed.add(path)
-            iterator.next()  # noqa: B305
-
-    return list(traversed)
